extracted ast helpers into separate class
