+++ src/types/ast.ts
@@ -1,0 +1,6 @@
+export type AstNode = {
+  children: AstNode[];
+  value: string;
+  type: string;
+  text: string;
+};
